fixups/service: Clear the display and wait 60 seconds before starting playback
Since there is no option for me to use a button to trigger a safe
shutdown of the Raspberry Pi (due to the e-paper display hat),
the only way to turn off the whole thing while the display is cleared
is to do an power off -> on sequence. Once the power is on and the Pi
has booted, the service will start and the display will be cleared,
there will be 60 seconds to turn off the whole device before it starts
video playback.

(I know that this is not safe, since this can, and most probably will,
cause filesystem errors in the future.)