slow-movie-player-service: Use ImageMagick for image dithering because I failed miserably
I am keeping the previous commit as well because someone might read that
and this one day.

My implementation of the Floyd-Steinberg dithering algorithm was painfully
slow. I have tried to optimize the algorithm using lookup tables, going full
numpy, etc. Every time the performance got worse... I have even tried to use
PIL/Pillow, but the results were abysmal.

In the end I went with using ImageMagick because it is so versatile and it is
very easy to supply image data for it via stdin and get the image data out from
stdout. The dithered images that ImageMagick produced were vastly superior to
my implementation or PIL/Pillow's, and the performance on a Raspberry Pi 3B+
is great.