slow-movie-player-service: Add timeout and debug info when update-display hangs
Sometimes the update-display process seems to get into an infinite loop or
wait because it hangs and it does not return.

This commit is supposed to prevent the process running for more than a specified
amount of time and add debug information (contents of /proc/PID/status,
contents of /proc/PID/stack if available and a full stack trace made with gdb).